requirements.txt
# AWS SDK
boto3==1.29.7

# Core ML dependencies
torch==2.5.1
torchvision==0.20.1
pillow>=9.0.0
numpy<2.0.0
# ultralytics>=8.3.0,<9.0.0
ultralytics==8.4.11
opencv-python-headless==4.10.0.84
pyyaml>=6.0

# Additional utilities
python-dotenv>=1.0.0
scikit-learn>=1.7.2

# FOR MACBOOK (only for local test)
# boto3==1.29.7
# torch>=2.0.0
# torchvision>=0.15.0
# pillow>=9.0.0
# numpy<2.0.0
# ultralytics>=8.3.0,<9.0.0
# opencv-python>=4.8.0
# pyyaml>=6.0
# python-dotenv>=1.0.0